This is all my fault. I'm sorry, everybody.
As 2401 said, because of some rules, I can not use the current method to sync public resources of PhoenixBuilder client.

I'm sorry that, all because of me, no one will ever see the day Omega Builder is open source.
I'm the bane of it.

The following is the change log.
* now we do not check the initialize counts of public resources of PhoenixBuilder client
* remove GetExtendOperation() from omega
* when omega is enabled, do not update the resources data for PhoenixBuilder itself
* remove my components named ChangeItemNameByUseAnvil and ExportMCStructure
* revert some changes which changed on bff8092 (InventoryContent related)